usm: dispatcher: Remove code duplication
The dispatcher will cache the skb_info and conn_tuple in a percpu array, each program will fetch it
and copy to a local instance. That change allows us to remove a lot of code duplication, and have a
single entry point rather than an entrypoint for prebuilt and another for runtime.

Made the change for http, http2 and kafka.